skip nonexisting files when copying deps
git-svn-id: https://dng.biozentrum.unibas.ch/svn/openstructure/trunk@2613 5a81b35b-ba03-0410-adc8-b2c5c5119f08

+++ deployment/macos/deps.py
@@ -151,6 +151,8 @@ def copy_deps(dependencies, outdir, use_rpath):
   for dep in dependencies:
     if dep.endswith('.dylib'):
       dst_name=os.path.join(outdir, 'lib', os.path.basename(dep))
+      if not os.path.exists(dep):
+        continue
       shutil.copy(dep, dst_name)
       if use_rpath:
         change_id(os.path.basename(dep), dst_name, use_rpath)
